Major changes in timing, handling the keypad and arrow keys, shift trap
There's still the issue of ultra-fast key-presses, but I'm afraid fixing
the Apple firmware deficiencies will only hinder the arrow keys
unusable.
I guess it's a different way the Macintosh and the Windows PCs are
using the arrow keys: selecting text on the Macs was probably considered
a mouse-only activity. Win PCs are able to use arrows and I guess the
mouse is no longer new and "cool" to use over the keyboard.