Fixed a couple of typos.
The main improvement is the addition of the M0110A keyboard in the
supported list. The code changes are somewhat significant.
Note: DO NOT USE THE KEYPAD =, /, * and + keys WITH SHIFT.
Not only do they not have any function, the Apple implementation of the
keyboard's firmware has made them problematic.
Also, there *might* be some glitches when using the arrow keys and shift
very quickly.
You have been warned.
Cheers!